+++ pytefas/__init__.py
@@ -20,7 +20,7 @@ from .exceptions import (
     TefasRateLimitError,
 )
 
-__version__ = "0.1.1"
+__version__ = "0.2.0"
 __all__ = [
     "Crawler",
     "TefasError",